requirements.txt
aiofiles==0.5.0
aniso8601==7.0.0       
async-exit-stack==1.0.1
async-generator==1.10  
backcall==0.2.0        
cachetools==4.2.1
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
colorama==0.4.4
DateTimeRange==1.0.0
decorator==4.4.2
dnspython==2.1.0
email-validator==1.1.2
fastapi==0.63.0
google-api-core==1.26.3
google-api-python-client==2.0.2
google-auth==1.28.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.4.4
googleapis-common-protos==1.53.0
graphene==2.1.8
graphql-core==2.3.2
graphql-relay==2.0.1
h11==0.12.0
httplib2==0.19.1
idna==2.10
ipykernel==5.5.3
ipython==7.22.0
ipython-genutils==0.2.0
itsdangerous==1.1.0
jedi==0.18.0
Jinja2==2.11.3
jupyter-client==6.1.12
jupyter-core==4.7.1
MarkupSafe==1.1.1
mbstrdecoder==1.0.1
oauthlib==3.1.0
orjson==3.5.1
packaging==20.9
parso==0.8.2
pickleshare==0.7.5
promise==2.3
prompt-toolkit==3.0.18
protobuf==3.15.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pydantic==1.8.1
Pygments==2.8.1
pyparsing==2.4.7
python-dateutil==2.8.1
python-dotenv==0.16.0
python-multipart==0.0.5
pytz==2021.1
PyYAML==5.4.1
pyzmq==22.0.3
requests==2.25.1
requests-oauthlib==1.3.0
rsa==4.7.2
Rx==1.6.1
six==1.15.0
starlette==0.13.6
tornado==6.1
traitlets==5.0.5
typepy==1.1.4
typing-extensions==3.7.4.3
ujson==3.2.0
uritemplate==3.0.1
urllib3==1.26.4
uvicorn==0.13.4
watchgod==0.7
wcwidth==0.2.5
websockets==8.1